requirements.txt
django
djangorestframework
mysqlclient

nodriver
beautifulsoup4
lxml
pydantic

sentry-sdk